Carry a caller's row identifier through the in-memory indexes as an int64
rtree_insert, sptree_insert, their temporal and split variants and the two
nearest-neighbour cursors take the identifier as an int, while a caller
identifies a row with an int64: DuckDB's row_t is int64, so a binding narrows
every identifier it stores and the index answers with a truncated one.

Widen the identifier to int64 through the whole path: the six insert entry
points, rtree_nn_cursor_next and sptree_nn_cursor_next, the ids array of an
RTreeNode and the id of an SPTree node, and every MeosArray that a search, a
join or a dedup fills. meos_array_add copies elem_size bytes, so an array
created at one width and read at another is silent rather than a compile error;
the element size and the cast move together at each of the sites that consume a
search result.

Collapse the duplicates of a split search by sorting rather than by a set
indexed with the identifier. A set sized to the largest identifier allocates in
proportion to a value the caller chooses, which reaches gigabytes on a sparse
identifier space and has no bound at 64 bits; sorting costs memory proportional
to the number of candidates.

+++ meos/include/meos.h
@@ -368,9 +368,9 @@ extern RTree *rtree_create_stbox();
 extern RTree *rtree_create_tpcbox();
 #endif
 extern void rtree_free(RTree *rtree);
-extern void rtree_insert(RTree *rtree, void *box, int id);
-extern void rtree_insert_temporal(RTree *rtree, const Temporal *temp, int id);
-extern void rtree_insert_temporal_split(RTree *rtree, const Temporal *temp, int id, int maxboxes);
+extern void rtree_insert(RTree *rtree, void *box, int64 id);
+extern void rtree_insert_temporal(RTree *rtree, const Temporal *temp, int64 id);
+extern void rtree_insert_temporal_split(RTree *rtree, const Temporal *temp, int64 id, int maxboxes);
 extern int rtree_search(const RTree *rtree, RTreeSearchOp op, const void *query, MeosArray *result);
 extern int rtree_join(const RTree *rtree1, const RTree *rtree2, RTreeSearchOp op, MeosArray *result);
 extern int rtree_search_temporal(const RTree *rtree, RTreeSearchOp op, const Temporal *temp, MeosArray *result);
@@ -382,7 +382,7 @@ extern int rtree_search_temporal_dedup(const RTree *rtree, RTreeSearchOp op, con
 typedef struct RTreeNNCursor RTreeNNCursor;
 
 extern RTreeNNCursor *rtree_nn_cursor_open(const RTree *rtree, const void *query);
-extern bool rtree_nn_cursor_next(RTreeNNCursor *cursor, int *id_out, double *dist_out);
+extern bool rtree_nn_cursor_next(RTreeNNCursor *cursor, int64 *id_out, double *dist_out);
 extern void rtree_nn_cursor_close(RTreeNNCursor *cursor);
 
 /**
@@ -413,9 +413,9 @@ extern SPTree *sptree_create_stbox(SPTreeKind kind);
 extern SPTree *sptree_create_tpcbox(SPTreeKind kind);
 #endif
 extern void sptree_free(SPTree *sptree);
-extern void sptree_insert(SPTree *sptree, void *box, int id);
-extern void sptree_insert_temporal(SPTree *sptree, const Temporal *temp, int id);
-extern void sptree_insert_temporal_split(SPTree *sptree, const Temporal *temp, int id, int maxboxes);
+extern void sptree_insert(SPTree *sptree, void *box, int64 id);
+extern void sptree_insert_temporal(SPTree *sptree, const Temporal *temp, int64 id);
+extern void sptree_insert_temporal_split(SPTree *sptree, const Temporal *temp, int64 id, int maxboxes);
 extern int sptree_search(const SPTree *sptree, RTreeSearchOp op, const void *query, MeosArray *result);
 extern int sptree_search_temporal(const SPTree *sptree, RTreeSearchOp op, const Temporal *temp, MeosArray *result);
 extern int sptree_search_temporal_dedup(const SPTree *sptree, RTreeSearchOp op, const Temporal *temp, int maxboxes, MeosArray *result);
@@ -426,7 +426,7 @@ extern int sptree_search_temporal_dedup(const SPTree *sptree, RTreeSearchOp op, 
 typedef struct SPNNCursor SPNNCursor;
 
 extern SPNNCursor *sptree_nn_cursor_open(const SPTree *sptree, const void *query);
-extern bool sptree_nn_cursor_next(SPNNCursor *cursor, int *id_out, double *dist_out);
+extern bool sptree_nn_cursor_next(SPNNCursor *cursor, int64 *id_out, double *dist_out);
 extern void sptree_nn_cursor_close(SPNNCursor *cursor);
 
 /*****************************************************************************
